Respond to thread requests in main thread
It's slow and lagging, and framerate depends on text busyness. Still,
the editor already works, running in a separate thread.

The need to constantly communicate text width is a problem and it
shouldn't be done over the channels. Currently it's not possible to
have font on thread side, or to ask rasterizer about text width.