User app error recovery
The `lovr.errhand` can be set to lite-lovr handler to enable error
recovery. It will pass the error message into the log view of each
editor instance, and run a LOVR loop that keeps editors functional.

This can be used to fix the error in user code. After it is fixed,
the LOVR environment needs a full restart to reload the user app
together with editors, which unfortunately drops lite context (editor
positions, editor layouts, opened files, undo stack...).